# eskip: sort built-in predicates