--- a/cmd/eskip/doc.go
+++ b/cmd/eskip/doc.go
@@ -83,7 +83,6 @@ const (
 	appendFiltersUsage  = "append filters to each patched route"
 	appendFileUsage     = "append filters from a file to each patched route"
 	prettyUsage         = "prints routes in a more readable format"
-	sortPredicateUsage  = "sort routes predicates"
 	indentStrUsage      = "indent string used in pretty printing. Must match regexp \\s"
 	jsonUsage           = "prints routes as JSON"